Moved contract various states to state.rs
