+++ chain-signatures/contract/src/state.rs
@@ -1,0 +1,44 @@
+use std::collections::HashSet;
+
+use borsh::{self, BorshDeserialize, BorshSerialize};
+use near_sdk::{AccountId, PublicKey};
+use serde::{Deserialize, Serialize};
+
+use crate::primitives::{Candidates, Participants, PkVotes, Votes};
+
+#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, Debug)]
+pub struct InitializingContractState {
+    pub candidates: Candidates,
+    pub threshold: usize,
+    pub pk_votes: PkVotes,
+}
+
+#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, Debug)]
+pub struct RunningContractState {
+    pub epoch: u64,
+    pub participants: Participants,
+    pub threshold: usize,
+    pub public_key: PublicKey,
+    pub candidates: Candidates,
+    pub join_votes: Votes,
+    pub leave_votes: Votes,
+}
+
+#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, Debug)]
+pub struct ResharingContractState {
+    pub old_epoch: u64,
+    pub old_participants: Participants,
+    // TODO: only store diff to save on storage
+    pub new_participants: Participants,
+    pub threshold: usize,
+    pub public_key: PublicKey,
+    pub finished_votes: HashSet<AccountId>,
+}
+
+#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, Debug)]
+pub enum ProtocolContractState {
+    NotInitialized,
+    Initializing(InitializingContractState),
+    Running(RunningContractState),
+    Resharing(ResharingContractState),
+}
